Location_Task_Alert
Location Task Manager is a location-based alarm clock which wakes you up
at the location of your choice. This is a simple and perfect App to
remind you the task to be performed once the user reaches the location
or left the location by setting a location alarm.
It allows storing multiple locations alarm, save locations to reuse them
later. This makes setting up a wake location quick and easy.
This app tracks the location changes on movement and triggers an alarm
with task details if current location matches with the stored location
tasks.
Task can be added by touching the location on the Google Map or entering
the location address manually.